Create lessons BatchController
Separate batch lesson creation from single-lesson creation

Add LessonCreation concern to handle shared params and logic
